backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# Celery & Redis
celery==5.3.6
redis==5.0.1

# S3/MinIO
minio==7.2.3
boto3==1.34.34

# Database Connectors
pyodbc==5.0.1
pymysql==1.1.0
cx-Oracle==8.3.0
# sqlalchemy connectors are above

# SFTP/FTP
paramiko==3.4.0

# Data Processing (for report execution)
pandas==2.1.4
lxml==5.1.0
openpyxl==3.1.2

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
croniter==2.0.1

# Validation
jsonschema==4.21.1

# Monitoring & Logging
structlog==24.1.0

# Testing (optional for dev)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # for testing
